Former-commit-id: 30d21f0d49d117f39b55e61d2c662dc86c07d3c4

+++ Control/xAODRootAccess/Root/THolder.cxx
@@ -15,12 +15,9 @@
 #include "xAODRootAccess/tools/THolder.h"
 #include "xAODRootAccess/tools/Message.h"
 #include "xAODRootAccess/tools/TDestructorRegistry.h"
+#include "THolderCache.h"
 
 namespace xAOD {
-
-   // Initialise the static member(s):
-   THolder::TypeCache_t THolder::s_typeMap = THolder::TypeCache_t();
-   THolder::SharedCount_t THolder::s_sharedCount = THolder::SharedCount_t();
 
    THolder::THolder()
       : m_object( 0 ), m_type( 0 ), m_typeInfo( 0 ), m_owner( kFALSE ) {
@@ -39,7 +36,7 @@ namespace xAOD {
 
       // Increase the object count:
       if( m_object && m_owner ) {
-         s_sharedCount[ m_object ]++;
+         Internal::THolderCache::instance().incRef( m_object );
       }
    }
 
@@ -48,7 +45,7 @@ namespace xAOD {
 
       // Increase the object count:
       if( m_object && m_owner ) {
-         s_sharedCount[ m_object ]++;
+         Internal::THolderCache::instance().incRef( m_object );
       }
    }
 
@@ -65,7 +62,7 @@ namespace xAOD {
 
       // Increase the object count:
       if( m_object && m_owner ) {
-         s_sharedCount[ m_object ]++;
+         Internal::THolderCache::instance().incRef( m_object );
       }
    }
 
@@ -91,12 +88,9 @@ namespace xAOD {
    THolder::~THolder() {
 
       // Delete the object using its dictionary:
-      if( m_object && m_owner ) {
-         s_sharedCount[ m_object ]--;
-         if( s_sharedCount[ m_object ] <= 0 ) {
-            s_sharedCount.erase( m_object );
-            deleteObject();
-         }
+      if( m_object && m_owner &&
+          ( ! Internal::THolderCache::instance().decRef( m_object ) ) ) {
+         deleteObject();
       }
    }
 
@@ -112,12 +106,9 @@ namespace xAOD {
       if( &rhs == this ) return *this;
 
       // Clean up the previously managed object:
-      if( m_object && m_owner ) {
-         s_sharedCount[ m_object ]--;
-         if( s_sharedCount[ m_object ] <= 0 ) {
-            s_sharedCount.erase( m_object );
-            deleteObject();
-         }
+      if( m_object && m_owner &&
+          ( ! Internal::THolderCache::instance().decRef( m_object ) ) ) {
+         deleteObject();
       }
 
       // Do the copy:
@@ -128,7 +119,7 @@ namespace xAOD {
 
       // Increase the object count:
       if( m_object && m_owner ) {
-         s_sharedCount[ m_object ]++;
+         Internal::THolderCache::instance().incRef( m_object );
       }
 
       // Return this same object:
@@ -150,12 +141,9 @@ namespace xAOD {
       if( &rhs == this ) return *this;
 
       // Clean up the previously managed object:
-      if( m_object && m_owner ) {
-         s_sharedCount[ m_object ]--;
-         if( s_sharedCount[ m_object ] <= 0 ) {
-            s_sharedCount.erase( m_object );
-            deleteObject();
-         }
+      if( m_object && m_owner &&
+          ( ! Internal::THolderCache::instance().decRef( m_object ) ) ) {
+         deleteObject();
       }
 
       // Do the copy:
@@ -189,12 +177,9 @@ namespace xAOD {
       if( m_object == obj ) return;
 
       // Delete the previous object:
-      if( m_object && m_owner ) {
-         s_sharedCount[ m_object ]--;
-         if( s_sharedCount[ m_object ] <= 0 ) {
-            s_sharedCount.erase( m_object );
-            deleteObject();
-         }
+      if( m_object && m_owner &&
+          ( ! Internal::THolderCache::instance().decRef( m_object ) ) ) {
+         deleteObject();
       }
 
       // Hold on to the new object from now on:
@@ -202,7 +187,7 @@ namespace xAOD {
 
       // Increase the object count:
       if( m_object && m_owner ) {
-         s_sharedCount[ m_object ]++;
+         Internal::THolderCache::instance().incRef( m_object );
       }
 
       return;
@@ -218,21 +203,18 @@ namespace xAOD {
       // Check if anything needs to be done:
       if( state == m_owner ) return;
 
-      if( m_object && m_owner ) {
-         s_sharedCount[ m_object ]--;
-         if( s_sharedCount[ m_object ] <= 0 ) {
-            ::Warning( "xAOD::THolder::setOwner",
-                       "Deleting object %p no longer held by any owner",
-                       m_object );
-            s_sharedCount.erase( m_object );
-            deleteObject();
-         }
+      if( m_object && m_owner &&
+          ( ! Internal::THolderCache::instance().decRef( m_object ) ) ) {
+         ::Warning( "xAOD::THolder::setOwner",
+                    "Deleting object %p no longer held by any owner",
+                    m_object );
+         deleteObject();
       }
 
       m_owner = state;
 
       if( m_object && m_owner ) {
-         s_sharedCount[ m_object ]++;
+         Internal::THolderCache::instance().incRef( m_object );
       }
 
       return;
@@ -271,19 +253,15 @@ namespace xAOD {
          return m_object;
       }
 
-      // Description of the class the user is asking for:
-      ::TClass* userClass = 0;
-
       // Check if we already know about this type:
-      TypeCache_t::const_iterator type_itr = s_typeMap.find( &tid );
-      if( type_itr == s_typeMap.end() ) {
-         // Try to access the dictionary for the requested type:
+      ::TClass* userClass = Internal::THolderCache::instance().getClass( tid );
+      // If not, look for it now:
+      if( ! userClass ) {
          userClass = ::TClass::GetClass( tid );
-         s_typeMap[ &tid ] = userClass;
-      } else {
-         // Get the dictionary from the internal cache:
-         userClass = type_itr->second;
+         Internal::THolderCache::instance().addClass( tid, userClass );
       }
+
+      // If we still don't have a dictionary, that's an issue:
       if( ! userClass ) {
          if( ! silent ) {
             ::Error( "xAOD::THolder::getAs",
@@ -345,18 +323,15 @@ namespace xAOD {
    void THolder::renew() {
 
       // Delete the object using its dictionary:
-      if( m_object && m_owner ) {
-         s_sharedCount[ m_object ]--;
-         if( s_sharedCount[ m_object ] <= 0 ) {
-            s_sharedCount.erase( m_object );
-            deleteObject();
-         }
+      if( m_object && m_owner &&
+          ( ! Internal::THolderCache::instance().decRef( m_object ) ) ) {
+         deleteObject();
       }
 
       // Create a new object:
       m_owner = kTRUE;
       m_object = m_type->New();
-      s_sharedCount[ m_object ]++;
+      Internal::THolderCache::instance().incRef( m_object );
 
       // Return gracefuly:
       return;
